+++ RELEASE_NOTES.md
@@ -2,6 +2,7 @@
 
 ### 0.7.0  Multi-Select (WIP)
 - #129 Fixes FileSystem Move and Copy to handle empty folders
+- #9 Changes Delete to enumerate and work on single items
 
 ### 0.6.2  (2022-12-23)
 - #121 Adds command to Open Root Directory